simbridge: Add "unaligned" string to unaligned io/mem reads and writes
There was an issue recently discovered that was due to an unaligned read
of a BAR register value. Specifically the ionic_dev_info_regs.fw_status
value, which is periodically read by the linux-ionic driver to check
the FW's status. This was causing all devcmd(s) to fail because before
each devcmd the fw_status is read and checked. The result of that
read is shown below:

simdevice_memrd(0xfe600007, 0x1) = 0x1000000

This change doesn't fix anything, but will cause the previous to now
be printed as:

simdevice_memrd(0xfe600007, 0x1) = 0x1000000 (unaligned)

Also, add the same warning to memwr, iord, and iowr.

Hopefully, if there are other issues related to unaligned reads/writes
in the future they will be easier to spot due to improved debug output.

Signed-off-by: Brett Creeley <brett.creeley@amd.com>